# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CTF Full V1 API
  version: 0.1.0
  description: >
    Partial contract skeleton for VS01-VS05 and the read-only External Reality
    Interface. All paths below are relative to the /api/v1 server URL. The runnable
    FastAPI application's generated /openapi.json is authoritative where this
    skeleton does not yet declare an implemented endpoint or exact response shape.
    Domain implementations must expand schemas without weakening ownership, version,
    evidence, genealogy, idempotency, or human-authority requirements. Every
    consequential mutation requires Idempotency-Key in the deployed policy and
    may return 413 or 429 with a stable error envelope.
...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /system/capabilities:
    get:
      tags: [System]
      security: []
      operationId: listSystemCapabilities
      description: Public, evidence-backed architecture capability inventory. No credentials or private runtime configuration are returned.
      parameters:
        - name: status
          in: query
          schema: {type: array, items: {type: string, enum: [IMPLEMENTED, PARTIAL, NOT_IMPLEMENTED, BLOCKED_EXTERNAL, DEFERRED_V1]}}
        - name: priority
          in: query
          schema: {type: array, items: {type: string, enum: [P0, P1, P2, P3]}}
      responses:
        "200":
          description: Filtered inventory with unfiltered public summary counts
          content:
            application/json:
              schema: {type: object}
        "503":
          description: Checked-in capability manifest is missing or invalid
  /system/readiness:
    get:
      tags: [System]
      security: []
      operationId: getSystemReadiness
      description: Safe aggregate of release blockers, AI, persistence, object storage, document worker, KHAL, and pilot readiness.
      responses:
        "200":
          description: Operator-safe readiness summary without secrets, URLs, or credentials
          content:
            application/json:
              schema: {type: object}
        "503":
          description: Checked-in capability manifest is missing or invalid
...
components:
  securitySchemes:
    bearerAuth: {type: apiKey, in: header, name: X-Session-Token}
    providerCredential: {type: apiKey, in: header, name: X-Provider-Credential}
...